+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project (tries to) adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.83] - 2026-07-24
+
+- [debug correlation steps](https://github.com/ACE-Collective/ace3/pull/413)
+- [Makes ACE more resilient to transient Splunk 401 auth errors](https://github.com/ACE-Collective/ace3/pull/412)
+
+## [3.0.82] - 2026-07-22
+
+- [Instrument the hunt → alert submission path](https://github.com/ACE-Collective/ace3/pull/411)
+- [surface transactionId from hunter to alert](https://github.com/ACE-Collective/ace3/pull/409)
+
+## [3.0.81] - 2026-07-22
+
+- [Adds jQuery support to JS deobfuscator](https://github.com/ACE-Collective/ace3/pull/407)
+- [Extracts several more email headers as observables](https://github.com/ACE-Collective/ace3/pull/406)
+
 ## [3.0.80] - 2026-07-20
 
 - [Cleans up alert mode changes](https://github.com/ACE-Collective/ace3/pull/402)